convert double to float in division

+++ src/Models/PurchasedItem.java
@@ -4,7 +4,7 @@ public class PurchasedItem {
 	public int ID;
 	public String name;
 	public String description;
-	public double quantity;
-	public double cost;
+	public float quantity;
+	public float cost;
 	public String date;
 }
